Moved some things around in the backend module and ripped out the tests there as they weren't actually testing anything
